refactor: do not export gitdiff
